STYLE : log 추가

+++ src/main/java/HeoJin/crawling_spring/okitchen/service/OkitchenService.java
@@ -100,7 +100,7 @@ public class OkitchenService {
                 continue;
             }
             cookingOrders.add(CookingOrder.builder()
-                    .step((Integer) cookingOrderStep)
+                    .step(cookingOrderStep)
                     .instruction(order.text())
                     .build());
             log.info("Step: " + order.text());
